moves logged_exec from main.py to common.py
